Them chuc nag dang nhap admin co session

+++ Weblaptop/app/Http/Controllers/DanhmucsanphamController.php
@@ -25,7 +25,16 @@ class DanhmucsanphamController extends Controller
      */
     public function create()
     {
-        return view('admincp.danhmucsanpham.create');
+        //Session admin check
+        $admin_account  = Session()->get('admin_name');
+        if($admin_account){
+            return view('admincp.danhmucsanpham.create');
+        }
+        else{
+            return view('admincp.loginadmin');
+        }
+
+        
     }
 
     /**
@@ -36,27 +45,38 @@ class DanhmucsanphamController extends Controller
      */
     public function store(Request $request)
     {
-        $data = $request->validate([
-            'tendanhmuc' => 'required|unique:danhmuc|max:255',
-            'slug_danhmuc' => 'required|unique:danhmuc|max:255',
-            'mota' => 'required|max:255',
-            'kichhoat' => 'required'
-        ],
-        [
-            'tendanhmuc.unique'=> 'Tên danh mục đã có rồi! Điền tên khác',
-            'tendanhmuc.unique'=> 'Slug danh mục đã có rồi! Điền slug khác',
-            'slug_danhmuc.required' => 'slug danh mục phải có',
-            'tendanhmuc.required' => 'Tên danh mục phải có',
-            'mota.required' => 'Mô tả phải có',
-        ]
-        );
-        $danhmucsanpham = new Danhmucsanpham();
-        $danhmucsanpham->tendanhmuc = $data['tendanhmuc'];
-        $danhmucsanpham->slug_danhmuc = $data['slug_danhmuc'];
-        $danhmucsanpham->mota = $data['mota'];
-        $danhmucsanpham->kichhoat = $data['kichhoat'];
-        $danhmucsanpham->save();
-        return redirect()->back()->with('status','Thêm danh mục sản phẩm thành công');
+
+        //Session admin check
+        $admin_account  = Session()->get('admin_name');
+        if($admin_account){
+            $data = $request->validate([
+                'tendanhmuc' => 'required|unique:danhmuc|max:255',
+                'slug_danhmuc' => 'required|unique:danhmuc|max:255',
+                'mota' => 'required|max:255',
+                'kichhoat' => 'required'
+            ],
+            [
+                'tendanhmuc.unique'=> 'Tên danh mục đã có rồi! Điền tên khác',
+                'tendanhmuc.unique'=> 'Slug danh mục đã có rồi! Điền slug khác',
+                'slug_danhmuc.required' => 'slug danh mục phải có',
+                'tendanhmuc.required' => 'Tên danh mục phải có',
+                'mota.required' => 'Mô tả phải có',
+            ]
+            );
+            $danhmucsanpham = new Danhmucsanpham();
+            $danhmucsanpham->tendanhmuc = $data['tendanhmuc'];
+            $danhmucsanpham->slug_danhmuc = $data['slug_danhmuc'];
+            $danhmucsanpham->mota = $data['mota'];
+            $danhmucsanpham->kichhoat = $data['kichhoat'];
+            $danhmucsanpham->save();
+            return redirect()->back()->with('status','Thêm danh mục sản phẩm thành công');
+        }
+        else{
+            return view('admincp.loginadmin');
+        }
+
+
+       
     }
 
     /**
@@ -76,9 +96,18 @@ class DanhmucsanphamController extends Controller
     // bị lỗi route đến destroy nên để vào đây dù gì cũng không dùng
     public function show($id)
     {
-        //xoa thu
-        Danhmucsanpham::find($id)->delete();
-        return redirect()->back()->with('status','Xoá danh mục sản phẩm thành công');
+        //Session admin check
+        $admin_account  = Session()->get('admin_name');
+        if($admin_account){
+            //xoa thu
+            Danhmucsanpham::find($id)->delete();
+            return redirect()->back()->with('status','Xoá danh mục sản phẩm thành công');
+        }
+        else{
+            return view('admincp.loginadmin');
+        }
+
+        
     }
 
     /**
@@ -90,8 +119,17 @@ class DanhmucsanphamController extends Controller
     public function edit($id)
     {
 
-        $danhmucsanpham = Danhmucsanpham::find($id);
-        return view('admincp.danhmucsanpham.edit')->with(compact('danhmucsanpham'));
+        //Session admin check
+        $admin_account  = Session()->get('admin_name');
+        if($admin_account){
+            //sua sp
+            $danhmucsanpham = Danhmucsanpham::find($id);
+            return view('admincp.danhmucsanpham.edit')->with(compact('danhmucsanpham'));
+        }
+        else{
+            return view('admincp.loginadmin');
+        }
+        
     }
 
     /**
@@ -103,25 +141,35 @@ class DanhmucsanphamController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $data = $request->validate([
+        
+        //Session admin check
+        $admin_account  = Session()->get('admin_name');
+        if($admin_account){
+            //cap nhat sp
+            $data = $request->validate([
             'tendanhmuc' => 'required|max:255',
             'slug_danhmuc' => 'required|max:255',
             'mota' => 'required|max:255',
             'kichhoat' => 'required'
-        ],
-        [
+            ],
+            [
             'tendanhmuc.required' => 'Tên danh mục phải có',
             'slug_danhmuc.required' => 'Tên danh mục phải có',
             'mota.required' => 'Mô tả phải có',
-        ]
-        );
-        $danhmucsanpham = Danhmucsanpham::find($id);
-        $danhmucsanpham->tendanhmuc = $data['tendanhmuc'];
-        $danhmucsanpham->slug_danhmuc = $data['slug_danhmuc'];
-        $danhmucsanpham->mota = $data['mota'];
-        $danhmucsanpham->kichhoat = $data['kichhoat'];
-        $danhmucsanpham->save();
-        return redirect()->back()->with('status','Cập nhật danh mục sản phẩm thành công');
+            ]
+            );
+            $danhmucsanpham = Danhmucsanpham::find($id);
+            $danhmucsanpham->tendanhmuc = $data['tendanhmuc'];
+            $danhmucsanpham->slug_danhmuc = $data['slug_danhmuc'];
+            $danhmucsanpham->mota = $data['mota'];
+            $danhmucsanpham->kichhoat = $data['kichhoat'];
+            $danhmucsanpham->save();
+            return redirect()->back()->with('status','Cập nhật danh mục sản phẩm thành công');
+        }
+        else{
+            return view('admincp.loginadmin');
+        }
+        
     }
 
     /**
@@ -132,7 +180,16 @@ class DanhmucsanphamController extends Controller
      */
     public function destroy($id)
     {
-        Danhmucsanpham::find($id)->delete();
-        return redirect()->back()->with('status','Xoá danh mục sản phẩm thành công');
+        //Session admin check
+        $admin_account  = Session()->get('admin_name');
+        if($admin_account){
+            // xoa sp
+            Danhmucsanpham::find($id)->delete();
+            return redirect()->back()->with('status','Xoá danh mục sản phẩm thành công');
+        }
+        else{
+            return view('admincp.loginadmin');
+        }
+        
     }
 }
